Added function to clear ledList hashMap and changed hashMap Logic
I made to fiarly significant changes in addition to several comments and
spacing updates.

1.  I created a function called clearHashMap  that is called in the
setup() function, it initializes the whole ledList to false.  I then
removed several lines of code that checked if the hashmap value was
null, as they should all be false.

2.  I created a new keypressed listener that calls clearHashMap() when
"DELETE" is pressed.  You can now output the hashmap then clear the cube
and start over.